requirements.txt
certifi==2025.4.26
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
